requirements.txt
alembic
boto3
kaggle
pandas
psycopg2-binary
